Allow commenting on issues that mention an old vscode version

+++ new-release/NewRelease.js
@@ -7,25 +7,33 @@ Object.defineProperty(exports, "__esModule", { value: true });
 exports.NewRelease = void 0;
 const utils_1 = require("../common/utils");
 class NewRelease {
-    constructor(github, label, labelColor, labelDescription, days) {
+    constructor(github, label, labelColor, labelDescription, days, oldVersionMessage) {
         this.github = github;
         this.label = label;
         this.labelColor = labelColor;
         this.labelDescription = labelDescription;
         this.days = days;
+        this.oldVersionMessage = oldVersionMessage;
     }
     async run() {
         const release = await (0, utils_1.loadLatestRelease)('stable');
         if (!(release && release.timestamp))
             throw Error('Could not load latest release');
         const daysSinceRelease = (Date.now() - release.timestamp) / (24 * 60 * 60 * 1000);
+        const issue = await this.github.getIssue();
+        const cleansed = issue.body.replace(/<!-- .* -->/g, '');
+        if (this.oldVersionMessage &&
+            !/VS ?Code Version:.*Insider/i.test(cleansed) &&
+            /VS ?Code Version:/i.test(cleansed) &&
+            !new RegExp(`VS ?Code Version:(.*[^\\d])?${release.productVersion.replace('.', '\\.')}([^\\d]|$)`, 'i').test(cleansed)) {
+            await this.github.postComment(this.oldVersionMessage.replace('{currentVersion}', release.productVersion));
+            return;
+        }
         if (daysSinceRelease > this.days) {
             // delete the label from the repo as a whole to remove it from all issues
             (0, utils_1.safeLog)('New release window passed. Globally deleting label ' + this.label);
             return this.github.deleteLabel(this.label);
         }
-        const issue = await this.github.getIssue();
-        const cleansed = issue.body.replace(/<!-- .* -->/g, '');
         if (!/VS ?Code Version:.*Insider/i.test(cleansed) &&
             new RegExp(`VS ?Code Version:(.*[^\\d])?${release.productVersion.replace('.', '\\.')}([^\\d]|$)`, 'i').test(cleansed)) {
             if (!(await this.github.repoHasLabel(this.label))) {
